+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 * Display application metrics using `mse status`
+* New option `--color` to enable/disable color for stdout/stderr
 
 ### Updated
 
@@ -11,6 +12,7 @@
 * `mse list` does not required the project name. It's optional.
 * Use `get_server_certificate` from `intel-sgx-ra`
 * Support refresh token expiration
+* Take the default configuration values from the backend 
 
 ## \[0.10.2\] - 2023-03-01